Refactoring: helper functions -> RamUtils
