+++ src/Makefile.am
@@ -101,7 +101,6 @@ souffle_sources = \
               GraphUtils.h                              \
               IODirectives.h                            \
               IOSystem.h                                \
-              RamIndexAnalysis.cpp  RamIndexAnalysis.h  \
               InlineRelationsTransformer.cpp            \
               LogStatement.h                            \
 	      InterpreterIndex.h            InterpreterIndex.cpp	\
@@ -112,7 +111,6 @@ souffle_sources = \
               PrecedenceGraph.cpp   PrecedenceGraph.h   \
               ProfileEvent.h                            \
               ProvenanceTransformer.cpp                 \
-              RamAnalysis.h                             \
 	      InterpreterContext.h                      \
 	      InterpreterEngine.cpp InterpreterEngine.h \
 	      InterpreterGenerator.h		        \
@@ -121,10 +119,13 @@ souffle_sources = \
 	      InterpreterProgInterface.h                \
 	      InterpreterPreamble.h			\
 	      InterpreterRecords.h         InterpreterRecords.cpp     \
+              RamAnalysis.h                             \
               RamComplexityAnalysis.cpp  RamComplexityAnalysis.h  \
-              RamLevelAnalysis.cpp  RamLevelAnalysis.h  \
               RamCondition.h                            \
+              RamExpression.h                           \
+              RamIndexAnalysis.cpp  RamIndexAnalysis.h  \
               RamNode.h                                 \
+              RamLevelAnalysis.cpp  RamLevelAnalysis.h  \
               RamOperation.h                            \
               RamProgram.h                              \
               RamRelation.h                             \
@@ -132,8 +133,9 @@ souffle_sources = \
               RamTransformer.cpp    RamTransformer.h    \
               RamTransforms.cpp     RamTransforms.h     \
               RamTranslationUnit.h                      \
-              RamExpression.h                           \
+              RamTypes.h                                \
               RamVisitor.h                              \
+              RamUtils.h                                \
               ReadStream.h                              \
               ReadStreamCSV.h                           \
               RelationRepresentation.h                  \
@@ -196,7 +198,6 @@ soufflepublic_HEADERS = \
                         PiggyList.h             \
                         ProfileDatabase.h       \
                         ProfileEvent.h          \
-                        RamTypes.h              \
                         ReadStream.h            \
                         ReadStreamCSV.h         \
                         SignalHandler.h         \
