+++ src/RamUtils.h
@@ -1,0 +1,95 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2020, The Souffle Developers. All rights reserved
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file RamUtils.h
+ *
+ * A collection of utilities operating on RAM constructs.
+ *
+ ***********************************************************************/
+
+#pragma once
+
+#include "RamCondition.h"
+#include "RamExpression.h"
+#include <memory>
+#include <vector>
+
+namespace souffle {
+
+/** @brief Determines if an expression represents an undefined value */
+inline bool isRamUndefValue(const RamExpression* expr) {
+    return nullptr != dynamic_cast<const RamUndefValue*>(expr);
+}
+
+/**
+ * @brief Convert terms of a conjunction to a list
+ * @param conds A RAM condition
+ * @return A list of RAM conditions
+ *
+ * Convert a condition of the format C1 /\ C2 /\ ... /\ Cn
+ * to a list {C1, C2, ..., Cn}.
+ */
+inline std::vector<std::unique_ptr<RamCondition>> toConjunctionList(const RamCondition* condition) {
+    std::vector<std::unique_ptr<RamCondition>> conditionList;
+    std::queue<const RamCondition*> conditionsToProcess;
+    if (condition != nullptr) {
+        conditionsToProcess.push(condition);
+        while (!conditionsToProcess.empty()) {
+            condition = conditionsToProcess.front();
+            conditionsToProcess.pop();
+            if (const auto* ramConj = dynamic_cast<const RamConjunction*>(condition)) {
+                conditionsToProcess.push(&ramConj->getLHS());
+                conditionsToProcess.push(&ramConj->getRHS());
+            } else {
+                conditionList.emplace_back(condition->clone());
+            }
+        }
+    }
+    return conditionList;
+}
+
+/**
+ * @brief Convert list of conditions to a conjunction
+ * @param A list of RAM conditions
+ * @param A RAM condition
+ *
+ * Convert a list {C1, C2, ..., Cn} to a condition of
+ * the format C1 /\ C2 /\ ... /\ Cn.
+ */
+inline std::unique_ptr<RamCondition> toCondition(const std::vector<const RamCondition*>& conds) {
+    std::unique_ptr<RamCondition> result;
+    for (const RamCondition* cur : conds) {
+        if (result == nullptr) {
+            result = std::unique_ptr<RamCondition>(cur->clone());
+        } else {
+            result = std::make_unique<RamConjunction>(
+                    std::move(result), std::unique_ptr<RamCondition>(cur->clone()));
+        }
+    }
+    return result;
+}
+
+/**
+ * @brief Convert list of conditions to a conjunction
+ * @param conds A list of RAM conditions
+ * @return A RAM condition
+ *
+ * Convert a list {C1, C2, ..., Cn} to a condition of
+ * the format C1 /\ C2 /\ ... /\ Cn.
+ */
+inline std::unique_ptr<RamCondition> toCondition(const std::vector<std::unique_ptr<RamCondition>>& conds) {
+    std::vector<const RamCondition*> args;
+    for (const auto& cur : conds) {
+        args.push_back(cur.get());
+    }
+    return toCondition(args);
+}
+
+}  // end of namespace souffle
